+++ BoardM5stackCore.cpp
@@ -4,8 +4,6 @@
 #include "BoardInterface.h"
 #include "Board320_240.h"
 #include "BoardM5stackCore.h"
-#include <SPI.h>
-#include <SD.h>
 
 /**
   Init board
@@ -19,41 +17,17 @@ void BoardM5stackCore::initBoard() {
   pinSpeaker = SPEAKER_PIN;
   pinBrightness = TFT_BL;
 
+  // Mute speaker
+  //ledcWriteTone(TONE_PIN_CHANNEL, 0);
+  digitalWrite(SPEAKER_PIN, 0);
+
+  //
   Board320_240::initBoard();
 }
 
 void BoardM5stackCore::mainLoop() {
 
   Board320_240::mainLoop();
-
-/*  if (!SD.begin(TFCARD_CS_PIN, SPI, 27000000)) {
-    Serial.println("Card Mount Failed");
-    return;
-  }
-  uint8_t cardType = SD.cardType();
-  if (cardType == CARD_NONE) {
-    Serial.println("No SD card attached");
-    return;
-  }
-
-  Serial.print("SD Card Type: ");
-  if (cardType == CARD_MMC) {
-    Serial.println("MMC");
-  } else if (cardType == CARD_SD) {
-    Serial.println("SDSC");
-  } else if (cardType == CARD_SDHC) {
-    Serial.println("SDHC");
-  } else {
-    Serial.println("UNKNOWN");
-  }
-
-  uint64_t cardSize = SD.cardSize() / (1024 * 1024);
-  Serial.printf("SD Card Size: %lluMB\n", cardSize);
-  
-  Serial.printf("Total space: %lluMB\n", SD.totalBytes() / (1024 * 1024));
-  Serial.printf("Used space: %lluMB\n", SD.usedBytes() / (1024 * 1024));  
-  */
 }
-
 
 #endif // BOARDM5STACKCORE_CPP